ConcordiaItemView: avoid API subrequest
Calling the API from inside a view opens up a number of 
potential problems (app instances may not have complete 
internet access to public API endpoints, the code will block
if the web server is close to maximum capacity and queues 
requests, etc.) and has a substantial performance penalty.

This changes the requests call to simply load the item and
serialize it directly so the data is in the same format as 
the API but without the encode/transfer/decode overhead.

This pattern will need to be applied throughout the other
views but long term it would probably be better to simply
enable django-rest-framework’s built-in template renderer
for any view which is supposed to have this 1:1 data
correspondence.